Implement initial DSSE verification
This adds dsse envelop processing and tests. This also adds an
inmemory keystore that can be used for now until we figure out
what we are using long term for fetching and storing public keys
for verification. This also creates a verifier interface along
with a Sigstore verifier for the purposes of verifying DSSE
envelopes through Sigstore

Signed-off-by: Michael Lieberman <mlieberman85@gmail.com>