Viper config implementation (#276)
* Consolidate duplicate flag declarations

A bit of uneeded DRY. Both commands share the same set of flags
so consolidate into the root init. If individual flags are ever
needed, we can set them in the init of each cmd.

Signed-off-by: Luke Hinds <lhinds@redhat.com>

* Implement viper for configuration mgmt

This change allows three types of input (in
order of load priority), in accordance with the
12 factor app design.

* CLI flags
* Environment Variables
* Configuration file

Environment variables are declared using the `GUAC`
prefix, e.g..

`export GUAC_GDBUSER=neo4j`

It requires the former creds value be split over two flags

`--gdbuser neo4j --gdbpass password`

Signed-off-by: Luke Hinds <lhinds@redhat.com>

* Add Luke to TAM (#271)

Signed-off-by: Brandon Lum <lumjjb@gmail.com>

Signed-off-by: Brandon Lum <lumjjb@gmail.com>

* Consolidate duplicate flag declarations

A bit of uneeded DRY. Both commands share the same set of flags
so consolidate into the root init. If individual flags are ever
needed, we can set them in the init of each cmd.

Signed-off-by: Luke Hinds <lhinds@redhat.com>

Signed-off-by: Luke Hinds <lhinds@redhat.com>

* Fix goimport lint complaints

Signed-off-by: Luke Hinds <lhinds@redhat.com>

* Handle err on viper.BindPflag

Signed-off-by: Luke Hinds <lhinds@redhat.com>

* Final lint fix

Signed-off-by: Luke Hinds <lhinds@redhat.com>

Signed-off-by: Luke Hinds <lhinds@redhat.com>
Signed-off-by: Brandon Lum <lumjjb@gmail.com>
Co-authored-by: Brandon Lum <lumjjb@gmail.com>