Replaced returning err with fmt.Errorf
- Replaced the `return err` with `return fmt.Errorf()`

Returning `fmt.Errorf()` instead of `return err` in Go is generally considered to be best practice because it allows for more detailed and informative error messages. `fmt.Errorf()` is a function that formats a string and returns an error, which means that developers can include additional context and information in the error message that can be helpful for debugging and troubleshooting. In contrast, simply returning `err` without any additional context can make it more difficult to understand the cause of the error and how to fix it. Additionally, by returning a formatted error message, it makes error handling and debugging more consistent.

Signed-off-by: nathannaveen <42319948+nathannaveen@users.noreply.github.com>