[pkg ingestion] Handle qualifiers in neo4j backend
Pushing them into version as a list of strings where each odd element is
the key and followed by the value. This list is sorted by keys.

This ensures that we can keep the query explicitly written out, although
there is a lot of wrangling from the golang structs to the list used in
ingestion and back to the golang structs used on the return path.

Made sure that no duplicates are created by running the following
queries and analyzing the neo4j data in the browser:

```gql
fragment allPkgTree on Package {
  type
  namespaces {
    namespace
    names {
      name
      versions {
        version
        qualifiers {
          key
          value
        }
        subpath
      }
    }
  }
}

mutation PkgM1 {
  ingestPackage(pkg: {type: "pypi", name: "tensorflow"}) {
    ...allPkgTree
  }
}

mutation PkgM2 {
  ingestPackage(
    pkg: {type: "pypi", name: "tensorflow", qualifiers: [{key: "arch", value: "amd64"}, {key: "distro", value: "stretch"}]}
  ) {
    ...allPkgTree
  }
}

mutation PkgM3 {
  ingestPackage(pkg: {type: "pypi", name: "tensorflow", version: "2.12.0"}) {
    ...allPkgTree
  }
}

mutation PkgM4 {
  ingestPackage(
    pkg: {type: "pypi", name: "tensorflow", version: "2.12.0", qualifiers: [{key: "arch", value: "amd64"}, {key: "distro", value: "stretch"}]}
  ) {
    ...allPkgTree
  }
}

mutation PkgM5 {
  ingestPackage(
    pkg: {type: "pypi", name: "tensorflow", version: "2.12.0", qualifiers: [{key: "distro", value: "stretch"}, {key: "arch", value: "amd64"}]}
  ) {
    ...allPkgTree
  }
}

mutation PkgM6 {
  ingestPackage(
    pkg: {type: "pypi", name: "tensorflow", version: "2.12.0", subpath: "foo", qualifiers: [{key: "distro", value: "stretch"}, {key: "arch", value: "amd64"}]}
  ) {
    ...allPkgTree
  }
}

mutation PkgM7 {
  ingestPackage(
    pkg: {type: "pypi", name: "tensorflow", version: "2.12.0", subpath: "foo", qualifiers: [{key: "distro", value: "stretch"}, {key: "arch", value: "amd64"}]}
  ) {
    namespaces {
      names {
        versions {
          qualifiers {
            key
            value
          }
        }
      }
    }
  }
}
```

Last mutation only retrieves the qualifiers part of the JSON. Since the
Cipher query still needs to build all the path to the last node (in case
it does not exist it needs to be created), there is no need to do extra
work to find the fields specified on the return path of the GraphQL
query and edit the Cipher query accordingly.

Signed-off-by: Mihai Maruseac <mihaimaruseac@google.com>