Split `CertifyScorecard`, separating subject from object.
This is because the ingestion will require these to be separate,
creating the object only when certification exists.

For now, I did not do the same split for the input spec, but it should
be doable if we want a consistent interface, though it will make the
query JSON nest one extra level.

Signed-off-by: Mihai Maruseac <mihaimaruseac@google.com>