Split CertifyScorecard, separating subject from object. (#502)
* Remove `mutation.gql` as it is included in `packages.gql`

Signed-off-by: Mihai Maruseac <mihaimaruseac@google.com>

* Fix some empty lines

Signed-off-by: Mihai Maruseac <mihaimaruseac@google.com>

* Separate backend interface functions

Signed-off-by: Mihai Maruseac <mihaimaruseac@google.com>

* Use `scorecards` for the query instead of `CertifyScorecard`

Since `CertifyScorecard` is a verb, using it for mutation instead.

While here, also fix formatting, naming (start lowercase as that's the
GraphQL convention), comments.

Includes the generated code too, to ensure that code still builds.

Signed-off-by: Mihai Maruseac <mihaimaruseac@google.com>

* Split `CertifyScorecard`, separating subject from object.

This is because the ingestion will require these to be separate,
creating the object only when certification exists.

For now, I did not do the same split for the input spec, but it should
be doable if we want a consistent interface, though it will make the
query JSON nest one extra level.

Signed-off-by: Mihai Maruseac <mihaimaruseac@google.com>

* Fix format

Signed-off-by: Mihai Maruseac <mihaimaruseac@google.com>

---------

Signed-off-by: Mihai Maruseac <mihaimaruseac@google.com>