[in-memory][backend] Fix index out of range (#532)
* [in-memory][backend] Fix index out of range

When ingesting packages in the in-memory backend, we assumed qualifiers
are given by `"{key}={value}"` strings. But this is not the case in the
neo4j backend or when doing the ingestion correctly via the graphql
client.

The `{key}={value}` assumption results in an index ouf of bounds issue
when ingesting from the graphql client.

Signed-off-by: Mihai Maruseac <mihaimaruseac@google.com>

* Fix format

Signed-off-by: Mihai Maruseac <mihaimaruseac@google.com>

---------

Signed-off-by: Mihai Maruseac <mihaimaruseac@google.com>