Update artifacts, hashEquals, and isOccurrance with internal structs and ids (#600)
* Update artifacts, hashEquals, and isOccurrance with internal structs and ids.

These GQL objects have IDs that are returned and supported in queries. Also
backedges are kept in internal structs from the noun object to the verb
objects. Ingestion and querying is reworked to use these internal structs.

Signed-off-by: Jeff Mendoza <jlm@jlm.name>

* Some review comments, use pkgNameorVersion interface.

Signed-off-by: Jeff Mendoza <jlm@jlm.name>

* Review comments for ingestion and ID matching.

Comment out the "registerAll" functions and all test data ingestion will be
handled by the client, ex: `cmd/graphql_playground/cmd/ingest.go`.

When a query has an ID provided, return the matching object if found, ignoring
all other fields in the spec.

Signed-off-by: Jeff Mendoza <jlm@jlm.name>

* Helper getters and setters. Cleanup pkg version name vs version.

Add get and set for hashEquals on artifacts and occurrences on artifacts,
packages, and sources. Remove isDependency and occurrences on pkg name as these
are only supposed to be on pkg version. Clean up pkgNameOrVersion interface to
reflect that.

Signed-off-by: Jeff Mendoza <jlm@jlm.name>

* Use nodeID() to convert uint32 to string.

Signed-off-by: Jeff Mendoza <jlm@jlm.name>

---------

Signed-off-by: Jeff Mendoza <jlm@jlm.name>