Backend cleanup and rename
Rename the testing backend to inmem, so that it does not clash with the Go
testing framework. Also rename the neo4j backend to match its directory
name. Make the required functions capital in the node interface. Organize the
imports properly. Remove the MaxUint usage. Use the generics-based
slices.Sort() where appropriate.

Signed-off-by: Jeff Mendoza <jlm@jlm.name>