Improve query performance for verbs / links. (#636)
If a subject or object is specified on a verb query that exactly maps to a
single node (ex: ID or full details are specified), then only search the
backedges of that node for any matching verbs. Code cleanup on these functions
for consistancy, and ease of copying. Change byid functions to use a single one
with generic types. Standardise on using ParseUint() instead of Atoi().

Signed-off-by: Jeff Mendoza <jlm@jlm.name>