Log and continue on ingest errors with bulk assemler. (#1275)
When ingesting, continue on error. Currently the incomplete backends panic when
calling mutations that are not implemented yet. We still want `guacone collect
files` to ingest all the nouns/verbs that are implemented. This is a short term
fix, as we want to differentiate between actual errors and not-implemented
apis in the future.

Signed-off-by: Jeff Mendoza <jlm@jlm.name>