Add SECURITY-INSIGHTS
Security insights (https://github.com/ossf/security-insights-spec)
has released as 1.0 and projects are starting to adopt it. It is
a low cost, low upkeep way to point people in the right direction
of various security related aspects of the GUAC project like where
our SBOMs are located, what security steps we take in our CI
process, etc. It also is useful as an indicator for anything we
might be missing and might need improvement. In the future most
if not all of this file might be able to be autogenerated, e.g.
where the SBOMS are for our released but for now it seems not too
bad to maintain it manually

Signed-off-by: Michael Lieberman <mlieberman85@gmail.com>