Change Keys method in kv interface to Scan (#1558)
Keys was a placeholder, as an external keyvalue may have more keys than we want
to read at once. Additionally we don't want to overload the external store with
an operation that will block for a long time. Now the Scan interface will only
get some number of keys at a time. Implementations for Redis and TiKV have been
added.

Signed-off-by: Jeff Mendoza <jlm@jlm.name>