Group golang dependabot updates
This should reduce some CI pressure, at the risk of increasing the
PR size under review.

For PRs that update graphql/atlas/any library that then requires
changing generated code, we would need to run `make generate` on top of
the large diff. But, even this should be less frequent now, since
instead of having to do this for 1 PR that changes gql and 1 PR that
changes atlas we can do both in the same PR. If the diff size becomes
too large we could exclude these, just like we do with the golang/x
ones.

Signed-off-by: Mihai Maruseac <mihaimaruseac@google.com>