Ungroup updates to go dependencies (#2656)
In #2613 I tried to group all dependencies in an attempt to reduce the number of times we need to rerun CI when it is flaky.

Unfortunately, this causes an issue since now the big PR that gets created for all the dependencies would just fail all CI jobs and have a large diff.

Meanwhile, with the split of CI intro multiple workflows, the flakiness decreased, so the problem is not that big, I think.

Signed-off-by: Mihai Maruseac <mihaimaruseac@google.com>